move `client.py` into `local`
